Made changes according to errors and db structure

+++ src/main/java/com/cbs/cbsclass/controller/TransactionController.java
@@ -28,5 +28,5 @@ public class TransactionController {
 class Tbody {
     long accountno_s;
     long accountno_r;
-    double amount;
+    float amount;
 }
